added new controller base class
